# openapi.json
{
  "components": {
    "schemas": {
      "ApplyCost": {
        "description": "What an Operator is agreeing to when they Apply.",
        "properties": {
          "queries_at_risk": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Queries running or queued on the Cluster right now, every one of which a Rollout terminates. A live reading and not a promise: it is already stale, and it is null when the Cluster could not be asked or when no restart is coming.",
            "title": "Queries At Risk"
          },
          "restarts_coordinator": {
            "description": "Whether applying the Candidate as it stands requires a Rollout.",
            "title": "Restarts Coordinator",
            "type": "boolean"
          },
          "warning": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "What to show an Operator before they Apply.",
            "title": "Warning"
          }
        },
        "required": [
          "restarts_coordinator"
        ],
        "title": "ApplyCost",
        "type": "object"
      },
...
      "Catalog": {
        "description": "A Catalog as it is stored and returned.",
        "properties": {
          "connector": {
            "pattern": "^[a-z][a-z0-9_-]{0,62}$",
            "title": "Connector",
            "type": "string"
          },
          "name": {
            "pattern": "^[a-z][a-z0-9_]{0,62}$",
            "title": "Name",
            "type": "string"
          },
          "properties": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Properties",
            "type": "object"
          },
          "supported": {
            "description": "False when the connector has no curated schema. Its properties passed through unvalidated and are checked only by Trino at Apply.",
            "title": "Supported",
            "type": "boolean"
          }
        },
        "required": [
          "name",
          "connector",
          "supported"
        ],
        "title": "Catalog",
        "type": "object"
      },
      "CatalogUpdate": {
        "additionalProperties": false,
        "description": "A partial update. The name is immutable; renaming is a drop and a create,\nwhich is a different and more destructive operation.",
        "properties": {
          "connector": {
            "anyOf": [
              {
                "pattern": "^[a-z][a-z0-9_-]{0,62}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Connector"
          },
          "properties": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Properties"
          }
        },
        "title": "CatalogUpdate",
        "type": "object"
      },
      "CatalogWrite": {
        "additionalProperties": false,
        "description": "What an Operator sends.",
        "properties": {
          "connector": {
            "pattern": "^[a-z][a-z0-9_-]{0,62}$",
            "title": "Connector",
            "type": "string"
          },
          "name": {
            "pattern": "^[a-z][a-z0-9_]{0,62}$",
            "title": "Name",
            "type": "string"
          },
          "properties": {
            "additionalProperties": true,
            "title": "Properties",
            "type": "object"
          }
        },
        "required": [
          "name",
          "connector"
        ],
        "title": "CatalogWrite",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RevertEffect": {
        "description": "What was staged, and what an Apply would then do to the Cluster.\n\nThe catalog lists are the point of this model. A revert is not a paper change:\napplying it issues real DROP CATALOG statements, and \"3 catalogs will be dropped\"\nis a materially different warning from \"the configuration will be rewritten\".",
        "properties": {
          "catalogs_created": {
            "items": {
              "type": "string"
            },
            "title": "Catalogs Created",
            "type": "array"
          },
          "catalogs_dropped": {
            "description": "Catalogs an Apply will DROP. Destructive.",
            "items": {
              "type": "string"
            },
            "title": "Catalogs Dropped",
            "type": "array"
          },
          "catalogs_replaced": {
            "items": {
              "type": "string"
            },
            "title": "Catalogs Replaced",
            "type": "array"
          },
          "cost": {
            "$ref": "#/components/schemas/ApplyCost",
            "description": "What applying this would cost the Cluster. A recovery action that restarts Trino must say so as loudly as the change that made it necessary."
          },
          "from_snapshot": {
            "title": "From Snapshot",
            "type": "integer"
          },
          "other_sections_stay_at": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "The Snapshot every other Section remains at. Null for a Full Rollback.",
            "title": "Other Sections Stay At"
          },
          "sections": {
            "description": "The Sections replaced in the Candidate.",
            "items": {
              "enum": [
                "catalogs",
                "client_certificates",
                "certificate_mapping",
                "permissions",
                "resource_groups",
                "event_listeners"
              ],
              "type": "string"
            },
            "title": "Sections",
            "type": "array"
          },
          "summary": {
            "description": "What this does, in words, for an Operator to read.",
            "title": "Summary",
            "type": "string"
          }
        },
        "required": [
          "from_snapshot",
          "sections",
          "cost",
          "summary"
        ],
        "title": "RevertEffect",
        "type": "object"
      },
      "RevertRequest": {
        "additionalProperties": false,
        "description": "Which Snapshot to go back to. Shared by both recovery actions so the two cannot\ndrift into different shapes for the same question.",
        "properties": {
          "snapshot": {
            "title": "Snapshot",
            "type": "integer"
          }
        },
        "required": [
          "snapshot"
        ],
        "title": "RevertRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "summary": "Control plane for configuring one Trino cluster",
    "title": "Apchi",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/candidate/rollback": {
      "post": {
        "description": "Full Rollback: for a serious mistake, and never implicit.\n\nLike a Section Revert this only stages. The Snapshot being restored from is not\nmodified -- applying this produces a new one, so history is never rewritten.",
        "operationId": "rollback_api_v1_candidate_rollback_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevertRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevertEffect"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stage an earlier Snapshot in its entirety",
        "tags": [
          "candidate"
        ]
      }
    },
    "/api/v1/catalogs": {
      "get": {
        "operationId": "list_catalogs_api_v1_catalogs_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Catalog"
                  },
                  "title": "Response List Catalogs Api V1 Catalogs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List staged Catalogs",
        "tags": [
          "catalogs"
        ]
      },
      "post": {
        "operationId": "create_catalog_api_v1_catalogs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CatalogWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Catalog"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stage a new Catalog",
        "tags": [
          "catalogs"
        ]
      }
    },
    "/api/v1/catalogs/revert": {
      "post": {
        "description": "Stages into the Candidate; it does not apply. An ordinary POST /applies follows,\nlike any other edit.\n\nThe response says which catalogs an Apply would drop, and that reverting one Section\nwhile the others stay put produces a configuration that has never run.",
        "operationId": "revert_api_v1_catalogs_revert_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevertRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevertEffect"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stage this Section's content from an earlier Snapshot",
        "tags": [
          "catalogs"
        ]
      }
    },
    "/api/v1/catalogs/{name}": {
      "delete": {
        "operationId": "delete_catalog_api_v1_catalogs__name__delete",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Remove a staged Catalog",
        "tags": [
          "catalogs"
        ]
      },
      "get": {
        "operationId": "get_catalog_api_v1_catalogs__name__get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Catalog"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Fetch a staged Catalog",
        "tags": [
          "catalogs"
        ]
      },
      "patch": {
        "operationId": "update_catalog_api_v1_catalogs__name__patch",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CatalogUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Catalog"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Edit a staged Catalog",
        "tags": [
          "catalogs"
        ]
      }
    },
...
  }
}